+++ modules/graphics/functions/colormaplist.m
@@ -1,0 +1,19 @@
+%=============================================================================
+% Copyright (c) 2016-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+function varargout = colormaplist(varargin)
+  % List available colormaps
+  narginchk(0, 0);
+  nargoutchk(0, 1);
+  maps = string(ls([modulepath('graphics', 'functions'), '/colormaps/*.m']));
+  maps = replace(maps,".m", "");
+  maps = deblank(maps);
+  varargout{1} = sort(maps);
+end
+%=============================================================================