+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `scatter3`: 3-D scatter plot.
 - `WindowState` property to Figure objects.
+- `nebula` colormap.
+- `colormaplist` List available colormaps.
 - `arrayfun` apply function to each element of array.
 - Support for `nix develop`, enabling a reproducible Bash shell preconfigured with Nelson’s build environment.
   Refer to [BUILDING.md](./BUILDING.md) for detailed usage instructions.